requirements.txt
flake8==6.0.0
pytest==7.4.0
selenium==4.10.0
webdriver-manager==3.9.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
requests==2.31.0
Faker==18.13.0
